+++ src/main/java/cart/dto/request/ProductRequestDto.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.request;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.PositiveOrZero;
